merge

+++ docker/build.sh
@@ -18,13 +18,17 @@
 # under the License.
 
 #
-# Execute command within a docker container
+# Build a docker container and optionally execute command within a it
 #
-# Usage: build.sh <CONTAINER_TYPE> [--dockerfile <DOCKERFILE_PATH>] [-it]
-#                [--net=host] [--cache-from <IMAGE_NAME>] <COMMAND>
+# Usage: build.sh <CONTAINER_TYPE> [--tag <DOCKER_IMAGE_TAG>]
+#                [--dockerfile <DOCKERFILE_PATH>] [-it]
+#                [--net=host] [--cache-from <IMAGE_NAME>] [<COMMAND>]
 #
-# CONTAINER_TYPE: Type of the docker container used the run the build: e.g.,
-#                 (cpu | gpu)
+# CONTAINER_TYPE: Type of the docker container used the run the build,
+#                 e.g. "ci_cpu", "ci_gpu"
+#
+# DOCKER_IMAGE_TAG: (Optional) Docker image tag to be built and used.
+#                   Defaults to 'latest', as it is the default Docker tag.
 #
 # DOCKERFILE_PATH: (Optional) Path to the Dockerfile used for docker build.  If
 #                  this optional value is not supplied (via the --dockerfile
@@ -33,7 +37,7 @@
 # IMAGE_NAME: An image to be as a source for cached layers when building the
 #             Docker image requested.
 #
-# COMMAND: Command to be executed in the docker container
+# COMMAND (optional): Command to be executed in the docker container
 #
 SCRIPT_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 
@@ -44,6 +48,12 @@ shift 1
 # Dockerfile to be used in docker build
 DOCKERFILE_PATH="${SCRIPT_DIR}/Dockerfile.${CONTAINER_TYPE}"
 DOCKER_CONTEXT_PATH="${SCRIPT_DIR}"
+
+if [[ "$1" == "--tag" ]]; then
+    DOCKER_IMAGE_TAG="$2"
+    echo "Using custom Docker tag: ${DOCKER_IMAGE_TAG}"
+    shift 2
+fi
 
 if [[ "$1" == "--dockerfile" ]]; then
     DOCKERFILE_PATH="$2"
@@ -80,12 +90,12 @@ fi
 COMMAND=("$@")
 
 # Validate command line arguments.
-if [ "$#" -lt 1 ] || [ ! -e "${SCRIPT_DIR}/Dockerfile.${CONTAINER_TYPE}" ]; then
+if [ ! -e "${SCRIPT_DIR}/Dockerfile.${CONTAINER_TYPE}" ]; then
     supported_container_types=$( ls -1 ${SCRIPT_DIR}/Dockerfile.* | \
         sed -n 's/.*Dockerfile\.\([^\/]*\)/\1/p' | tr '\n' ' ' )
       echo "Usage: $(basename $0) CONTAINER_TYPE COMMAND"
       echo "       CONTAINER_TYPE can be one of [${supported_container_types}]"
-      echo "       COMMAND is a command (with arguments) to run inside"
+      echo "       COMMAND (optional) is a command (with arguments) to run inside"
       echo "               the container."
       exit 1
 fi
@@ -114,6 +124,7 @@ function upsearch () {
 # reasonable defaults if you run it outside of Jenkins.
 WORKSPACE="${WORKSPACE:-${SCRIPT_DIR}/../}"
 BUILD_TAG="${BUILD_TAG:-tvm}"
+DOCKER_IMAGE_TAG="${DOCKER_IMAGE_TAG:-latest}"
 
 # Determine the docker image name
 DOCKER_IMG_NAME="${BUILD_TAG}.${CONTAINER_TYPE}"
@@ -125,6 +136,9 @@ DOCKER_IMG_NAME=$(echo "${DOCKER_IMG_NAME}" | sed -e 's/=/_/g' -e 's/,/-/g')
 # Convert to all lower-case, as per requirement of Docker image names
 DOCKER_IMG_NAME=$(echo "${DOCKER_IMG_NAME}" | tr '[:upper:]' '[:lower:]')
 
+# Compose the full image spec with "name:tag" e.g. "tvm.ci_cpu:v0.03"
+DOCKER_IMG_SPEC="${DOCKER_IMG_NAME}:${DOCKER_IMAGE_TAG}"
+
 # Print arguments.
 echo "WORKSPACE: ${WORKSPACE}"
 echo "CI_DOCKER_EXTRA_PARAMS: ${CI_DOCKER_EXTRA_PARAMS[@]}"
@@ -132,12 +146,14 @@ echo "COMMAND: ${COMMAND[@]}"
 echo "CONTAINER_TYPE: ${CONTAINER_TYPE}"
 echo "BUILD_TAG: ${BUILD_TAG}"
 echo "DOCKER CONTAINER NAME: ${DOCKER_IMG_NAME}"
+echo "DOCKER_IMAGE_TAG: ${DOCKER_IMAGE_TAG}"
+echo "DOCKER_IMG_SPEC: ${DOCKER_IMG_SPEC}"
 echo ""
 
 
 # Build the docker container.
 echo "Building container (${DOCKER_IMG_NAME})..."
-docker build -t ${DOCKER_IMG_NAME} \
+docker build -t ${DOCKER_IMG_SPEC} \
     -f "${DOCKERFILE_PATH}" \
     ${CI_DOCKER_BUILD_EXTRA_PARAMS[@]} \
     "${DOCKER_CONTEXT_PATH}"
@@ -148,25 +164,29 @@ if [[ $? != "0" ]]; then
     exit 1
 fi
 
-# Run the command inside the container.
-echo "Running '${COMMAND[@]}' inside ${DOCKER_IMG_NAME}..."
+if [[ -n ${COMMAND} ]]; then
 
-# By default we cleanup - remove the container once it finish running (--rm)
-# and share the PID namespace (--pid=host) so the process inside does not have
-# pid 1 and SIGKILL is propagated to the process inside (jenkins can kill it).
-echo ${DOCKER_BINARY}
-${DOCKER_BINARY} run --rm --pid=host \
-    -v ${WORKSPACE}:/workspace \
-    -w /workspace \
-    -e "CI_BUILD_HOME=/workspace" \
-    -e "CI_BUILD_USER=$(id -u -n)" \
-    -e "CI_BUILD_UID=$(id -u)" \
-    -e "CI_BUILD_GROUP=$(id -g -n)" \
-    -e "CI_BUILD_GID=$(id -g)" \
-    -e "CI_PYTEST_ADD_OPTIONS=$CI_PYTEST_ADD_OPTIONS" \
-    -e "CI_IMAGE_NAME=${DOCKER_IMAGE_NAME}" \
-    ${CUDA_ENV}\
-    ${CI_DOCKER_EXTRA_PARAMS[@]} \
-    ${DOCKER_IMG_NAME} \
-    bash --login docker/with_the_same_user \
-    ${COMMAND[@]}
+    # Run the command inside the container.
+    echo "Running '${COMMAND[@]}' inside ${DOCKER_IMG_SPEC}..."
+
+    # By default we cleanup - remove the container once it finish running (--rm)
+    # and share the PID namespace (--pid=host) so the process inside does not have
+    # pid 1 and SIGKILL is propagated to the process inside (jenkins can kill it).
+    echo ${DOCKER_BINARY}
+    ${DOCKER_BINARY} run --rm --pid=host \
+        -v ${WORKSPACE}:/workspace \
+        -w /workspace \
+        -e "CI_BUILD_HOME=/workspace" \
+        -e "CI_BUILD_USER=$(id -u -n)" \
+        -e "CI_BUILD_UID=$(id -u)" \
+        -e "CI_BUILD_GROUP=$(id -g -n)" \
+        -e "CI_BUILD_GID=$(id -g)" \
+        -e "CI_PYTEST_ADD_OPTIONS=$CI_PYTEST_ADD_OPTIONS" \
+        -e "CI_IMAGE_NAME=${DOCKER_IMAGE_NAME}" \
+        ${CUDA_ENV}\
+        ${CI_DOCKER_EXTRA_PARAMS[@]} \
+        ${DOCKER_IMG_SPEC} \
+        bash --login docker/with_the_same_user \
+        ${COMMAND[@]}
+
+fi
